requirements.txt
gradio
pymongo
openai
python-dotenv
PyMuPDF
